Adjust battery script to report from sysfs
Reading values out of /proc/apm/... is not very reliable on many modern
laptops. Instead read the values out of sysfs.

Also drop support for ibam as it only reads from /proc/apm/. If sysfs
support is added to it in the future it should be reintroduced.

With the information exposed in sysfs we are able to calculate the
current power usage in terms of Watts, so report that. The remaining
time reported only considders the current since the unit of the battery
charge is uA.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>