Change lock keybindings
Ctrl+Alt+L has been in my finger memory for a long time, but it has
never really fit in with the rest of the wmii keybindings, and now
clashes with some keybindings I want to use on tmux, so it is time for
it to go!

We now use Mod+grave to lock the screen and Mod+shift+grave to toggle
automatic locking. This is going to be massive brainfuck for a while,
but I think it's for the best.

Users who still like the old Ctrl+Alt+L to lock can change their
wmiirc_local.py to call lock.registerTraditionalKeyBindings instead, or
define the bindings themselves if they weren't already.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>