xterm LINES / COLUMNS bug partial fix
It seems that xterm sets up it's own environment if it is empty when it
is started, but in doing so it sets up the LINES and COLUMNS environment
variables.

Although xterm does update these variables on the fly as it is resized,
curses applications will take the values as set when they were started
as gospel and will not respond to resize events.

If we pass the environment to xterm when spawning it it does not set up
it's own environment, does not set LINES or COLUMNS in the environment,
and curses applications will therefore respond to resize events.

This patch only fixes terminals spawned from a plugin, etc. I really
need to patch the call function in pygmi to fix this everywhere. Namely,
terminals spawned with Win+Enter still exhibit the bug.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>