Dell local: Add Intuos pad button bindings
For the moment just set the default bindings to mimic the defaults in
Windows. I'll no doubt customise this later to suit my workflow better.

I'd like one of the buttons toggle an on-screen keyboard, I think that
would be pretty useful for editing short text fields rather than
reaching over the top of the tablet to the keyboard.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>