Refactor wacom code into plugin
This takes the dell laptop specific wacom mapping code and turns it into
a generic plugin. Mappings should be defined in this plugin and the
local scripts should select the appropriate mapping.

As a bonus, this could be used to simplify using key bindings to swap
profiles for different applications.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>