Suppress duplicate exception notifications
I use notify_exception all over the place to try to make sure I see any
errors that do occur. The problem is, since notify_exception keeps
passing the exception up the stack, multiple notify_exception wrappers
can see the same exception and send multiple notifications about it.

This patch changes it such that only the topmost notify_exception
wrapper will send a notification, but the exception is still passed up
the stack to allow it to be handled at higher layers or printed on the
console.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>