xdg autostart compliance; persistant settings
This will parse .desktop files in ~/.config/autostart and
/etc/xdg/autostart and launch anything there following the spec.

plugins.local has the opportunity to fake our environment or force
specific autostart files to load, for the case where OnlyShowIn is set
and does not include WMII, yet the user does want it loaded.

This isn't 100% compliant with the xdg spec - in particular I'm relying
on the shell to handle quoted arguments in the Exec string and don't
expand %U, etc. but it's close enough to work for all the autostart
scripts I care about at the moment.

This also introduces pluginmanager.persist to track whether autostart
has been run even if the plugin is reloaded. I think it might be a good
idea to move other settings into persist as well to avoid the problem of
reloading a plugin resetting something overridden in local to default.

Note that I've only tested dynamically loading this plugin and haven't
actually tested logging in with this in place, so I haven't verified
that everything loads in the right order.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>