Don't retag clients on /./ (temp workaround for witray retagging)
witray spawns on tag '/./', i.e. all tags. However, the lock plugin
makes sure that nothing is ever added to the '!lock' tag and retags it
elsewhere, which somehow caused witray to end up on a single tag.

For now, don't retag anything tagged with /./ to avoid this. Ideally I
still want them not shown on !lock, and retagged with /./ when !lock is
destroyed.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>