Fix lock notifications not being removed
When lock.py was run manually (or triggered by xautolock), we didn't
wait for it to clean up it's notifications and they were being left
behind.

This sets daemon=False on the timer that cleans up these messages so
they will be cleaned.

In the long term I'd like to try using python-xlib to replace the need
for xautolock and triggering the lock script externally to remove this
uglyness.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>