Avoid runaway timers in music status
It seems evertime a Monitor.active is set a new timer is started, even
if the monitor already had an active timer. This can lead to more and
more CPU usage and since the timer ticks take a lock this quickly leads
to a huge amount of lock contention slowing down the user interface.

This is easy to reproduce and test:
1. wmiir xwrite /ctl exec wmii -r python/wmiirc
2. open htop and note the CPU usage of wmiirc
3. Start cmus playing some music
4. Hold mod+Vol Down to lower cmus volume to minimum, then release and
   hold mod+Vol Up to max it out. Repeat several times.
5. Release the keys!
6. Wait a few seconds for htop to update...
7. Note the new much higher average CPU usage.
8. Optionally strace and marvel at the lock contention.

Really, this is a bug in pygmi, but that's no excuse not to implement
the below trivial workaround.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>