Make volume keys intelligent
This patch makes the volume up/down keys intelligent by default. That
is, if a music player is running & currently playing music, the volume
command will be sent to it; otherwise the command will be sent to the
default mixer.

The default mixer now be adjusted by holding down the mod key while
adjusting the volume. Note that this essentially reverses the default
behaviour when a music player is running.

Perhaps I should add an option to disable this?

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>