Workaround some issues with xtrlock and keys held down
I recently noticed that if I held the lock combo down for too long that
the system would not actually lock - it changed to the !lock tag, but
then changed back without locking.

Similarly, I found that I could cause the current tag to continue being
displayed while the system was locked if I pressed the lock combo twice
in rapid succession.

In terms of the first issue, this patch removes the grab that wmii has
on all the keys before locking and increases the delay before spawning
xtrlock (to give the user more time to remove their fingers from the
lock combo, which would still prevent xtrlock from launching properly
even though we no longer have a grab on it). This should reduce the
chance of accidentally leaving the system unlocked by holding down the
keys for too long.

I went to fix the second issue by using a lock to prevent _locknow
running if it was already running, however after disabling the grab on
all the keys for the first issue I realised that the chance of this
second issue occuring became remote since the lock key could no longer
be pressed twice in rapid succession. As such this patch now only fixes
this issue as a side effect of the fixing the first issue.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>