Add shortcuts to release grab on XF86Audio keys
These can be used to allow other applications to use these keys if
desired.

I actually did this for spotify, only to discover that it doesn't
actually grab the keys (ie, it has to be the focussed window). I've
spotted that spoitfy has a d-bus interface, so I'll just hook that up to
the music plugin like I have cmus and mocp.

Still, this may be useful for other things, so here it is.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>