Split out dbus hacks into separate plugin
Spotify was going to be the third user of dbus, and with one main loop
each the hack was clearly going too far.

This splits out the dbus functionality into a separate plugin, so that
we can create exactly one thread to handle the dbus main loop (ideally
I'd integrate that main loop into the wmii event handler and skip the
extra thread, but 1. I'm not sure the wmii event loop is built with
extension in mind, and 2. I'd have to work out how the details of how
the glib dbus main loop actually works).

This also fixes a bug where we weren't initialising threading in glib
(apparently it requires two calls?) and the spotify plugin was causing
the interpreter to crash.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>