Add spotify plugin
This also adds another possible redirection for the volume keys since
calling the dbus methods to get/set the volume don't seem to do anything
for me and I want them redirected back to the general mixer. It's
possible that it's because I'm using bluetooth & pulseaudio, not sure -
that works fine for moc & cmus.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>