Fix dbus plugin unloading
This should make the purple plugin reloading clean, but the libnotify
unload function still needs some work. Namely I need to figure out how
to unregister the service from the bus - at the moment it is only
releasing the well known name, which isn't the same thing and will
interfere with reloading the plugin.

Note that reloading the wmiidbus plugin itself is still untested.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>