Action to change background + hook actions up to /event
This introduces the set_background action which takes a filename and
does the obvious thing. Running the plugin from the command line
dispatches an action to do the same thing.

More importantly this introduces a new plugin that matches 'Action'
events written to /event and dispatches them as actions. I'm not sure
why this isn't already hooked up as such by default in the python
version of wmiirc - it is in the shell version

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>