Spotify: Run from commandline to get artist & title
When one of these ads comes up I want to make sure I can easily get the
correct spelling for the blacklist. This patch allows running the
spotify plugin directly from the commandline to print out the artist and
title of the current track, even if spotify is paused at the time.

Note that because of some threading issues in pygmi an error will be
thrown while shutting down - this is normal (but would be nice to fix).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>