Mac Mini Local: Delay xmodmap after keyboard shortcut
Seems that xmodmap fails if it is run while certain keys are still held
down (despite the fact that it seems to recognise this situation and
explicitly prints out a message and delays for two seconds...),
preventing it from working properly when called from ctrl+alt+x (in
fact, it does enough to undo a previous run, but not enough to apply the
new modmap... sigh...).

Delay calling xmodmap to give enough time to release the keys.

A similar issue seems to happen when initially logging in, and the same
delay seems to work around that too. Not entirely sure what is going on
there.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>