Remove timer.join in wmiirc_local.py
The first time we log in, the scripts hang somewhere. gdb Shows it's
waiting on a semaphore, and my first guess is this timer.join(). We
don't really need it anyway, so see what happens if I remove it.

This does mean that the race I added this to avoid is re-opened, but
a) The other steps taken to avoid this will probably still minimise the
   chance of it happening, and
b) If it happens it's a pretty minor annoyance, whereas locking up is
   much more serious.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>