Replace buggy pygmi.call to prevent hangs
Finally got around to debugging why MythFrontend was hanging after
starting a video ~13 times - it was blocked writing to stdout. wmiirc on
the other end wasn't reading from the pipe, allowing the buffer to fill
up and blocking further writes.

The thing that most annoys me about this, is that you have to go out of
you way to open pipes with python's subprocess module, yet they still
managed to cock it up >-(

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>